backend/requirements.txt

aiohttp
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart
unstructured[pdf]>=0.10.30
unstructured-client
python-dotenv>=1.0.0
supabase>=2.0.3
pandas>=2.1.3
google-generativeai>=0.3.1
python-jose[cryptography]
pydantic>=2.5.2
pydantic-settings
pytest 
pytest-asyncio 
pytest-cov